Finish the type review: no sign conversions left in the core
167 locations down to none. -Wsign-conversion is clean across src/core and
src/persistence, checked under ASan over the goldens, the boundaries, the
templates, the epsilon-hull and the nine strategy variants: 54 tests, no
invalid access.

The three-kind rule held up and decided every site. Indices became size_t:
the frequency index in all five algorithms, the engine's and the 1D
union's index, the loop counters, digitCount, frequencyCount, pointCount,
MAXP, take, bucketSize, the seven specification slots. Counts stayed
signed, which is only linspace's N - the user's point count, validated on
entry and used as N - 1 for a divisor. And tagged selectors stayed signed
with an explicit cast where they index, which silences the warning and
says what is happening: algorithm MC's parameter, template_engine's slots
where slotOf returns -1, the termIndex that is -1 for the gain, and the
hull walk's b1, b2, nextPoint and previousPoint, which come from searches
returning -1 for "nowhere".

A fourth kind turned up at the end, in the contour tracer: grid
coordinates that do neighbour arithmetic in a Moore walk and are bounded
with "> 0" and "< width - 1". A negative intermediate is both possible and
caught there, and unsigned would turn an off-by-one from a negative index
into an enormous one. They stay signed.

Two things the review found that were not type warnings at all.

TemplateEngine's m_combinationCount was an int32 holding the PRODUCT of
the grid sizes. Eight uncertain parameters on a 25-point grid is 25^8,
about 1.5e11, which overflows a 32-bit int - and an overflowed count does
not make the sweep slow, it makes it silently wrong, reserving from
garbage and bounding the loop with it.

And the three expression builders looped "i < sizeNum - 1". With the empty
case guarded four lines above they were safe, but making sizeNum unsigned
changed the failure from benign, no iterations, to catastrophic, an
enormous bound. They read "i + 1 < sizeNum" now, which is correct without
depending on a distant if.

+++ src/core/boundaries/boundary_union_1d.cpp
@@ -25,7 +25,7 @@ std::int32_t BoundaryUnion1D::bucketIndex(double x, double totalPhase)
     return (std::int32_t) res;
 }
 
-void BoundaryUnion1D::insertSorted(TraceSet & layerBucketsRow, std::int32_t index, qftbx::NicholsPoint point, double totalPhase)
+void BoundaryUnion1D::insertSorted(TraceSet & layerBucketsRow, std::size_t index, qftbx::NicholsPoint point, double totalPhase)
 {
     //The iterator used to be passed in from the caller, computed BEFORE this
     //function might have grown the same bucket; taking the index alone and
@@ -38,14 +38,14 @@ void BoundaryUnion1D::insertSorted(TraceSet & layerBucketsRow, std::int32_t inde
         }
     }
 
-    bucket.insert(bucket.begin() + index, point);
+    bucket.insert(bucket.begin() + static_cast<std::ptrdiff_t>(index), point);
 }
 
 std::vector<TraceSet> BoundaryUnion1D::buildLayerBuckets(const TraceSet & chosenCurves, double totalPhase, bool open, bool upper)
 {
     std::vector<TraceSet> layerBuckets (kLayerCount);
 
-    for (std::int32_t i = 0; i < kLayerCount; i++)
+    for (std::size_t i = 0; i < kLayerCount; i++)
     {
         TraceSet & row = layerBuckets[i];
         row.resize(totalPhase + 1);
@@ -54,7 +54,7 @@ std::vector<TraceSet> BoundaryUnion1D::buildLayerBuckets(const TraceSet & chosen
         {
             Trace & bucket = row.at(static_cast<std::size_t>(bucketIndex(point.phase, totalPhase)));
 
-            std::int32_t index = 0;
+            std::size_t index = 0;
             for (const qftbx::NicholsPoint & placed : bucket)
             {
                 if (upper)
@@ -279,7 +279,7 @@ inline std::int32_t BoundaryUnion1D::bucketIndex(double x, double totalPhase, st
     return (std::int32_t) res;
 }
 
-TraceSet BoundaryUnion1D::buildUnionBuckets(const Trace & unionPoints, double totalPhase, std::int32_t pointCount)
+TraceSet BoundaryUnion1D::buildUnionBuckets(const Trace & unionPoints, double totalPhase, std::size_t pointCount)
 {
     TraceSet unionBucketsRow (pointCount);
 
@@ -404,7 +404,7 @@ void BoundaryUnion1D::run(const BoundaryData * boundaries, const TraceMetadata &
             }
         }
 
-        m_unionBuckets.push_back(buildUnionBuckets(unionPoints, totalPhase, phasePointCount));
+        m_unionBuckets.push_back(buildUnionBuckets(unionPoints, totalPhase, static_cast<std::size_t>(phasePointCount)));
 
         //Ordered by proximity only where the boundary is open: a closed one
         //is already a cycle.
